Add mongo database
- Separate all componenet
- store register data in mongo database
- add error handling and create common error handler
