+++ backend/controllers/imageController.js
@@ -1,12 +1,13 @@
 const tmp = require('tmp')
 const path = require('path')
 const fs = require('fs')
+const { successResponse, errorResponse } = require('../utils/errorHandler')
+const { HTTP_BAD_REQUEST_400, HTTP_INTERNAL_SERVER_ERROR_500 } = require('../utils/https.status')
 const { get_image_results, filter_results, find_most_likely_username, isValidUrl, prepare_display_data } = require('../services/imageService')
 
 const uploadImage = async (req, res) => {
-    console.log('here to show image', req.file)
     if (!req.file) {
-        return res.status(400).json({ error: 'No image file provided' });
+        return errorResponse(res, { error: 'No image file provided' }, 'No image file provided', HTTP_BAD_REQUEST_400)
     }
 
     try {
@@ -25,43 +26,41 @@ const uploadImage = async (req, res) => {
             fs.unlinkSync(filePath); // Delete the file
             fs.rmdirSync(tempDir.name); // Remove the temporary directory
 
-            return res.status(200).json({ results: results });
+            return successResponse(res, { results: results }, 'Image uploaded successfully')
         } catch (error) {
             // Clean up on error
-            fs.unlinkSync(filePath);
-            fs.rmdirSync(tempDir.name);
-            return res.status(500).json({ error: error.message });
+            fs.unlinkSync(filePath)
+            fs.rmdirSync(tempDir.name)
+            return errorResponse(res, { error: error.message }, error.message, HTTP_INTERNAL_SERVER_ERROR_500)
         }
 
     } catch (error) {
-        return res.status(500).json({ error: error.message });
+        return errorResponse(res, { error: error.message }, error.message, HTTP_INTERNAL_SERVER_ERROR_500)
     }
 }
 
 const filterResultsEndpoint = (req, res) => {
     const data = req.body;
     if (!data) {
-        return res.status(400).json({ error: 'No results provided in the request' });
+        return errorResponse(res, { error: 'No results provided in the request' }, 'No results provided in the request', HTTP_BAD_REQUEST_400)
     }
-    console.log(req.body)
     results = data['results']
 
-    const { instagram_usernames, linkedin_usernames, twitter_usernames, facebook_usernames, other_usernames } = filter_results(results)
+    const { instagramUsernames, linkedinUsernames, facebookUsernames, twitterUsernames, otherUsernames } = filter_results(results)
 
-    return res.status(200).json({
-        'instagram': instagram_usernames,
-        'linkedin': linkedin_usernames,
-        'twitter': twitter_usernames,
-        'facebook': facebook_usernames,
-        'others': other_usernames
-    });
+    return successResponse(res, {
+        'instagram': instagramUsernames,
+        'linkedin': linkedinUsernames,
+        'twitter': facebookUsernames,
+        'facebook': twitterUsernames,
+        'others': otherUsernames
+    }, 'Filter Result Successfully')
 }
 
-const prepareDisplayDataEndpoint = (req, res) => {
+const prepareDisplayDataEndpoint = async (req, res) => {
     const data = req.body;
-
     if (!data) {
-        return res.status(400).json({ error: 'Invalid request data' });
+        return errorResponse(res, { error: 'Invalid request data' }, 'Invalid request data', HTTP_BAD_REQUEST_400)
     }
 
     try {
@@ -71,17 +70,16 @@ const prepareDisplayDataEndpoint = (req, res) => {
         const facebookUsernames = data.facebook || [];
         const otherUsernames = data.others || [];
 
-        const displayData = prepare_display_data(
+        const displayData = await prepare_display_data(
             instagramUsernames,
             linkedinUsernames,
             twitterUsernames,
             facebookUsernames,
             otherUsernames
-        );
-
-        return res.status(200).json({ display_data: displayData });
+        )
+        return successResponse(res, { display_data: displayData }, 'Prepare display data successfully.')
     } catch (error) {
-        return res.status(500).json({ error: error.message });
+        return errorResponse(res, { error: error.message }, error.message, HTTP_INTERNAL_SERVER_ERROR_500)
     }
 }
 
@@ -89,7 +87,7 @@ const most_likely_username = (req, res) => {
     const data = req.body;
 
     if (!data) {
-        return res.status(400).json({ error: 'Invalid request data' });
+        return errorResponse(res, { error: 'Invalid request data' }, 'Invalid request data', HTTP_BAD_REQUEST_400)
     }
 
     try {
@@ -103,11 +101,10 @@ const most_likely_username = (req, res) => {
             facebookUsernames,
             twitterUsernames,
             linkedinUsernames
-        );
-
-        return res.status(200).json({ most_likely_usernames: mostLikely });
+        )
+        return successResponse(res, { most_likely_usernames: mostLikely }, 'Get the most likely username successfully.')
     } catch (error) {
-        return res.status(500).json({ error: error.message });
+        return errorResponse(res, { error: error.message }, error.message, HTTP_INTERNAL_SERVER_ERROR_500)
     }
 }
 
@@ -115,12 +112,12 @@ const validate_url = (req, res) => {
     const data = req.body;
 
     if (!data || !data.url) {
-        return res.status(400).json({ error: 'No URL provided' });
+        return errorResponse(res, { error: 'No URL provided' }, 'No URL provided', HTTP_BAD_REQUEST_400)
     }
 
     const url = data.url;
-    const isValid = isValidUrl(url);
-    return res.status(200).json({ url: url, is_valid: isValid });
+    const isValid = isValidUrl(url)
+    return successResponse(res, { url: url, is_valid: isValid }, 'Url validate successfully.')
 }
 
 module.exports = {
