+++ backend/package.json
@@ -11,9 +11,13 @@
   "license": "ISC",
   "dependencies": {
     "axios": "^1.7.9",
+    "bcryptjs": "^3.0.2",
     "cors": "^2.8.5",
+    "dotenv": "^16.4.7",
     "express": "^4.21.2",
     "fs": "^0.0.1-security",
+    "jsonwebtoken": "^9.0.2",
+    "mongoose": "^8.10.1",
     "multer": "^1.4.5-lts.1",
     "path": "^0.12.7",
     "readline": "^1.3.0",
