+++ backend/utils/https.status.js
@@ -1,0 +1,7 @@
+exports.HTTP_BAD_REQUEST_400 = 400
+exports.HTTP_UNAUTHORIZED_401 = 401
+exports.HTTP_INTERNAL_SERVER_ERROR_500 = 500
+exports.HTTP_RESOURCE_NOT_FOUND_404 = 404
+exports.HTTP_CONFLICT_409 = 409
+exports.HTTP_TOO_MANY_REQUESTS_429 = 429
+exports.HTTP_INVALID_CAPTCHA_430 = 430
